Attempt to fix issue #5.
Override the insync method to ignore 'unknown' as not being in sync if
ignore_impossible_operations is set to true.

This has a unit test, but I'm not sure it'll actually work (due to the way
puppet mangles values, we may get 'true' or :true instead of true..

Also ignore_impossible_operations is horribly named, suggestions welcome.

We should also have a version which throws an exception (i.e. fails, rather
than just doesn't converge).