Puppet 4 Support
Initial support for Puppet 4.  Removes errors thrown by the runtime as facter is
now a binary and doesn't have the libraries required by this module.  This removes
the dependency by performing all library routines in native ruby.  Obviously this
royally destroyed RSpec, so there is some fairly major refactoring done in order
to stub out the bits that need to be done for the existing tests.  Also included
some rudimentary tests for individual class methods.

CI for ruby 1.9.3 is broken hence the changes to the Gemfile, the upstream json
gems now require ruby ~> 2.