Puppet 4 Support
Initial support for Puppet 4.  Removes errors thrown by the runtime as facter is
now a binary and doesn't have the libraries required by this module.  This removes
the dependency by performing all library routines in native ruby.  The max interface
speed code looked suspect so I've cleaned that up and ensured it works on 16.04.